The most significant changes include the addition of new components and functionalities in various files.
In `SharedMenuSelector.razor`, a new unordered list has been added to display a list of menus. Each list item is clickable and triggers the `SelectMenu` function with the menu's id as a parameter. A new `SelectedId` parameter has also been added to the component.

`Dialog.razor` now contains a new div with a header that reads "This is a Test Dialog".

`TestPage.razor` has a new page with a button that opens a dialog when clicked, using the `DialogService`.

`FoodOptionsDialog.razor` has been significantly updated with a new component that displays information about a food item, including its name, description, price, and status. It also includes options to change the status of the food item to either "Out of Stock" or "Available". The component also displays a list of related active price lists.

`LanguageSelector.razor` now contains a new component that allows the user to select a language from a list of supported languages. The selected language is stored in the `StoreAppService`.

`Cart.razor` has a new page that displays information about a shopping cart for a specific table. The page includes information about the table, the status of the table, a QR code for the table, and a list of items in the shopping cart.

`StoreAppService.cs` has been updated to include additional functionality and data handling for the application. This includes the addition of new properties and methods for loading floor tables, updating food item status, getting option sets, loading price lists, and more.

`app.css` has been updated with new styles for various elements.

`blazor.boot.json` and `service-worker-assets.js` have been updated, likely as a result of the changes in the `StoreAppService.cs` file. The hash values for several files have been updated.

`MainLayout.razor` has seen the injection of the StoreAppService and the adjustment of the size of RadzenColumns. The text "Whitehorn Bistro 213-1" was replaced with a LanguageSelector component.

`Program.cs` has registered the SharedService for dependency injection and the StoreAppService was updated to include the SharedService.

`OrderDelightStoreApp.csproj` and `OrderDelightStoreApp.csproj.user` have seen changes in the reference to a folder and the path of the last used publish profile respectively.

`ActiveMenus.razor`, `Tables.razor`, and `FoodOptions.razor` have seen various changes to the layout and functionality of the menu, including changes to the way menu items are displayed and filtered, display of a list of tables from the StoreAppService, and removal of existing content respectively.

`OrderDelightStoreApp - Web Deploy.pubxml.user`, `OrderDelightStoreApp - Zip Deploy.pubxml.user`, and `OrderDelightStoreApp - Zip Deploy.pubxml` have seen updates in the history of deployments and the InstallAspNetCoreSiteExtension property set to false.

`App.razor` has seen the addition of RadzenDialog, RadzenNotification, RadzenTooltip, and RadzenContextMenu components.